Moved classes into different layers for organization
