+++ app/src/main/java/com/caruski/eatoutlog/domain/Dish.java
@@ -1,4 +1,4 @@
-package com.caruski.eatoutlog;
+package com.caruski.eatoutlog.domain;
 
 public class Dish {
