+++ app/src/main/res/layout/content_main.xml
@@ -9,7 +9,7 @@
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
     app:layout_behavior="@string/appbar_scrolling_view_behavior"
-    tools:context="com.caruski.eatoutlog.MainActivity"
+    tools:context="com.caruski.eatoutlog.activity.MainActivity"
     tools:showIn="@layout/activity_main">
 
 </RelativeLayout>
