+++ app/src/main/res/menu/main_context_menu.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context="com.caruski.eatoutlog.MainActivity">
+    tools:context="com.caruski.eatoutlog.activity.MainActivity">
 
     <item
         android:id="@+id/action_edit_restaurant"
