+++ app/src/main/res/menu/new_dish_image_context.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context="com.caruski.eatoutlog.NewDishActivity">
+    tools:context="com.caruski.eatoutlog.activity.NewDishActivity">
 
     <item
         android:id="@+id/action_gallery_image"
