+++ app/src/main/res/menu/view_dish_context_menu.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
 xmlns:tools="http://schemas.android.com/tools"
-tools:context="com.caruski.eatoutlog.ViewDishesActivity">
+tools:context="com.caruski.eatoutlog.activity.ViewDishesActivity">
 
 <item
     android:id="@+id/action_edit_dish"
